+++ src/index.ts
@@ -1,4 +1,3 @@
-export * from './source-map-support';
-export * from './esbuild';
-export * from './transform-dynamic-import';
+export * from './source-map';
+export * from './transform';
 export * from './resolve-ts-path';
